Minigrid_robust_ent/rl-starter-files/rl-starter-files/requirements.txt
torch-ac>=1.1.0
gym-minigrid
tensorboardX>=1.6
numpy>=1.3
scikit-image
gym>=0.18.0,<0.19.0